+++ components/Table/__tests__/index.spec.tsx
@@ -18,9 +18,9 @@ expect.extend(toHaveNoViolations);
 describe("DMP Table Component", () => {
   const columns = [
     { id: 'id', name: 'id', isRowHeader: false },
-    { id: 'name', name: 'Name Column', isRowHeader: true, allowsSorting: true },
-    { id: 'email', name: 'Email Column', isRowHeader: true },
-    { id: 'other', name: 'Other Column', isRowHeader: true },
+    { id: 'name', name: 'Name Column', isRowHeader: true, allowsSorting: true, direction: '' as const },
+    { id: 'email', name: 'Email Column', isRowHeader: true, direction: '' as const },
+    { id: 'other', name: 'Other Column', isRowHeader: true, direction: '' as const },
   ];
 
   const rows = Array.from({ length: 5 }, (_, i) => {
@@ -90,10 +90,10 @@ describe("DMP Table Component", () => {
 
     await waitFor(() => {
       const want = [
-        { id: 'id', name: 'id', isRowHeader: false },
+        { id: 'id', name: 'id', isRowHeader: false, direction: '' as const },
         { id: 'name', name: 'Name Column', isRowHeader: true, allowsSorting: true, direction: 'ascending' },
-        { id: 'email', name: 'Email Column', isRowHeader: true },
-        { id: 'other', name: 'Other Column', isRowHeader: true },
+        { id: 'email', name: 'Email Column', isRowHeader: true, direction: '' as const },
+        { id: 'other', name: 'Other Column', isRowHeader: true, direction: '' as const },
       ];
       expect(mockSort).toHaveBeenCalledWith(want);
       // TODO: Check that the correct icon is displayed icon
@@ -103,10 +103,10 @@ describe("DMP Table Component", () => {
     fireEvent.click(nameColumn);
     await waitFor(() => {
       const want = [
-        { id: 'id', name: 'id', isRowHeader: false },
+        { id: 'id', name: 'id', isRowHeader: false, direction: '' as const },
         { id: 'name', name: 'Name Column', isRowHeader: true, allowsSorting: true, direction: 'descending' },
-        { id: 'email', name: 'Email Column', isRowHeader: true },
-        { id: 'other', name: 'Other Column', isRowHeader: true },
+        { id: 'email', name: 'Email Column', isRowHeader: true, direction: '' as const },
+        { id: 'other', name: 'Other Column', isRowHeader: true, direction: '' as const },
       ];
       expect(mockSort).toHaveBeenCalledWith(want);
       // TODO: Check that the correct icon is displayed icon